requirements.txt
openai-whisper==20240930
PyAudio==0.2.14
sounddevice==0.5.1
pytest==8.3.4
